refactor: pkg/utils to validator and converter, redis function
